==> Notebooks/02 - Data cleaning and vaderSentiment analysis.ipynb <==
import html
import pandas as pd

from html.parser import HTMLParser
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# Helper functions for HTML clean-up.

class MLStripper(HTMLParser):
    def __init__(self):
        super().__init__()
        self.reset()
        self.fed = []
    def handle_data(self, d):
        self.fed.append(d)
    def get_data(self):
        return ' '.join(self.fed)

def strip_tags(html):
    s = MLStripper()
    s.feed(html)
    return s.get_data()

df = pd.read_csv('first_week_oct_2015_comments_by_top_400.csv')

df.shape

df.head()

df['text']  = df['text'].apply(html.unescape)

analyzer = SentimentIntensityAnalyzer()

def vaderize(sentence):
    return analyzer.polarity_scores(sentence)

df['scores'] = df['text'].apply(vaderize)

df.head()

df[['neg', 'neu', 'pos', 'compound']] = df.scores.apply(pd.Series)

for text in df.sort_values(by='neg', ascending=False)['text'].head(20):
    print(text, '\n')

df['text'] = df['text'].apply(strip_tags)

df.head()

df = df.drop('scores', axis=1)

df.head()

df.iloc[17]['text']

df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores.csv', index=False)
