==> Notebooks/03 - Textblob sentiment analysis.ipynb <==
import pandas as pd

from textblob import TextBlob

df = pd.read_csv('first_week_oct_2015_comments_by_top_400_with_scores.csv')

df.shape

df.head()

tb_sentiment = df['text'].apply(lambda text: TextBlob(text).sentiment)

tb_sentiment[0]

df[['tb_polarity', 'tb_subjectivity']] = tb_sentiment.apply(pd.Series)

df.head()

for text in df.sort_values(by='tb_polarity')['text'].head(20):
    print(text, '\n')

df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores_v2.csv', index=False)
