==> Notebooks/04 - PerspectiveAPI Sentiment Analysis.ipynb <==
import json
import requests
import numpy as np
import pandas as pd
import time

df = pd.read_csv('first_week_oct_2015_comments_by_top_400_with_scores_v2.csv')

df.shape

df.head()

# Add API key here.
api_key = ''

url = ('https://commentanalyzer.googleapis.com/v1alpha1/comments:analyze' +    
    '?key=' + api_key)

data_dict = {
    'comment': {'text': 'what kind of idiot name is foo?'},
    'languages': ['en'],
    'requestedAttributes': {'TOXICITY': {}}
}
response = requests.post(url=url, data=json.dumps(data_dict)) 
response_dict = json.loads(response.content) 
print(json.dumps(response_dict, indent=2))

def get_toxicity(comment):
    data_dict = {
        'comment': {'text': comment},
        'languages': ['en'],
        'requestedAttributes': {'TOXICITY': {}}
    }
    response = requests.post(url=url, data=json.dumps(data_dict)) 
    response_dict = json.loads(response.content)
    time.sleep(1)
    try:
        return response_dict['attributeScores']['TOXICITY']['summaryScore']['value']
    except:
        return np.nan

get_toxicity(df['text'][17])

df['toxicity'] = df['text'].apply(get_toxicity)

df.index[df['toxicity'].isna()]

get_toxicity(df['text'][19])

df.at[19, 'toxicity'] = get_toxicity(df['text'][19])

df.iloc[19]

df.at[39, 'toxicity'] = get_toxicity(df['text'][39])

df.iloc[39]

for text in df['text'][df['toxicity'].isna()]:
    print(text, '\n')

for comment in df.sort_values(by='toxicity', ascending=False)['text'].head(20):
    print(comment, '\n')

df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores_v3.csv', index=False)
