==> Notebooks/05 - Profanity check and feature engineering.ipynb <==
%%capture

# Import libraries.
import html
import pandas as pd

from html.parser import HTMLParser
from profanity_check import predict_prob

class tag_counter(HTMLParser):
    """
    Helper class for counting HTML tags.
    """
    def __init__(self):
        HTMLParser.__init__(self)
        self.num_tags = 0
    def handle_starttag(self, tag, attrs):
        self.num_tags +=1
    def count_tags(self, html_text):
        self.num_tags = 0
        self.feed(html_text)
        return self.num_tags

parser = tag_counter()

# Load primary dataframe.
df = pd.read_csv('first_week_oct_2015_comments_by_top_400_with_scores_v3.csv')

# Load original html comments from file.
html_text =  pd.read_csv('first_week_oct_2015_comments_by_top_400.csv', usecols = ['text'])

# Preview loaded data.
html_text.head()

# Remove excess whitespace from comment text.
df['text'] = df['text'].replace('\s+', ' ',regex=True)

# Add scores from profanity-check to dataframe.
df['pc_prob_offensive'] = predict_prob(df['text'])

# Examine comments identified as most likely to be offensive
# by profanity-check.
for comment in (df.sort_values(by='pc_prob_offensive', ascending=False)['text'].head(20)):
    print(comment, '\n')

# Restore original html comments to primary dataset.
df['html_text'] = html_text['text']

# Remove html escape sequences.
df['html_text'] = df['html_text'].apply(html.unescape)

# Check updated html text column.
df['html_text'].head()

# Test tag counting.
parser.count_tags(df['html_text'][17])

# Add tag count feature to dataframe.
df['num_tags'] = df['html_text'].apply(parser.count_tags)

# Preview distribution of tag counts.
df['num_tags'].value_counts()

# Add number of capital letters to dataframe.
df['num_caps'] = df['text'].apply(lambda text: 
                                  sum(1 for char in text if char.isupper()))

# Compare 'author' and 'by' fields.
# These are redundant for comments.
(df['author'] == df['by']).value_counts()

# Add text length feature to dataframe.
df['text_len'] = df['text'].str.len()

# Normalize capital letter count.
df['pct_caps'] = df['num_caps']/df['text_len']

# Examine comments with the highest proportion of capital 
# letters.
for comment in df.sort_values(by='pct_caps', ascending=False)['text'].head(20):
    print(comment, '\n')

# Normalize tag counts.
df['tags_per_char'] = df['num_tags']/df['text_len']

# Convert 'time_ts' column to pandas datetime.
# Overwrite Unix time (redundant).
df['time'] = pd.to_datetime(df['time_ts'])

# Extract hour from posting time.
df['hour_posted'] = df['time'].dt.hour

# Examine distribution of 'dead' values.
df.dead.value_counts()

# Fill NA values - probably safe to assume comments not 
# marked as 'dead' aren't.
df['dead'] = df['dead'].fillna(False)

# Check results.
df.dead.value_counts()

# Examine distribution of 'deleted' values.
df['deleted'].value_counts()

# Double-check results - nothing useful in this column!
df.deleted.isna().sum()

# Review updated column list.
df.columns

# Drop unneeded columns.
df = df.drop(['author', 
              'time_ts',
              'html_text', 
              'deleted',
              'num_caps', 
              'num_tags'], axis=1)

# Re-check column list
df.columns

# Reorder columns.
df = df[['id', 'parent', 'by', 'time', 'hour_posted', 'text',
         'dead', 'ranking', 'text_len', 'pct_caps',
         'tags_per_char', 
         'toxicity', 'neg', 'neu', 'pos', 'compound', 
         'tb_polarity', 'tb_subjectivity', 'pc_prob_offensive']]

# Re-name columns to highlight score sources.
df.columns = ['id', 'parent', 'by', 'time', 'hour_posted', 'text',
         'dead', 'ranking', 'text_len', 'pct_caps',
         'tags_per_char', 
         'papi_toxicity', 'v_neg', 'v_neu', 'v_pos', 'v_compound', 
         'tb_polarity', 'tb_subjectivity', 'pc_prob_offensive']

# Review updated dataframe.
df.head()

# Save to file.
df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores_and_features.csv', index=False)
