==> Notebooks/06 - TextBlob NaiveBayesAnalyzer.ipynb <==
# Import libraries.
import pandas as pd

from textblob import TextBlob
from textblob.sentiments import NaiveBayesAnalyzer

# Load data.
df = pd.read_csv('first_week_oct_2015_comments_by_top_400_with_scores_and_features.csv')

# Preview data.
df.head()

i = 0

def analyze(text):
    """
    TextBlob's NaiveBayesAnalyzer is an NLTK classifer trained on a movie 
    reviews corpus. It returns a namedtuple of the form Sentiment(classification, 
    p_pos, p_neg).
    """
    blob = TextBlob(text, analyzer=NaiveBayesAnalyzer())
    
    global i
    if i % 100 == 0:
        print(i)
    i+= 1
    
    return blob.sentiment

tb_nb_sentiment = df['text'].apply(analyze)

# Save to dataframe and to file (overnight run).
df['tb_nb_sentiment'] = tb_nb_sentiment
df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores_and_features_v2.csv',
         index=False)

# Sample output.
df['tb_nb_sentiment'].head()

# Extract probability of negative sentiment from naive Bayes output.
df['tb_nb_prob_neg'] = df['tb_nb_sentiment'].apply(lambda x: x[2])

# Drop original output; fields are redundant. (Probabilities sum to one.)
df = df.drop('tb_nb_sentiment', axis=1)

df.columns

# Re-order columns to group TextBlob output.
df = df[['id', 'parent', 'by', 'time', 'hour_posted', 'text', 'dead', 'ranking',
       'text_len', 'pct_caps', 'tags_per_char', 
       'papi_toxicity', 'v_neg', 'v_neu', 'v_pos', 'v_compound', 'tb_polarity',
       'tb_subjectivity', 'tb_nb_prob_neg', 'pc_prob_offensive']]

# Look at dataframe.
df.head()

# Examine 20 most surely negative comments by this metric.
for text in df.sort_values(by='tb_nb_prob_neg', ascending=False)['text'].head(20):
    print(text, '\n')

# Re-save to disk.
df.to_csv('first_week_oct_2015_comments_by_top_400_with_scores_and_features_v2.csv,
          index=False')
