==> Notebooks/07 - Baseline Models.ipynb <==
# Import libraries.
import numpy as np
import pandas as pd
import string

from sklearn.feature_selection import f_regression, RFE, SelectKBest
from sklearn.linear_model import LinearRegression, LogisticRegression
from sklearn.metrics import accuracy_score, mean_absolute_error, precision_score, recall_score
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler

# Load data.
df = pd.read_csv('first_week_oct_2015_comments_by_top_400_with_scores_and_features_v2.csv')

# Preview data.
df.head()

def avg_word_len(text):
    words = text.split()
    return sum(len(word) for word in 
               filter(lambda word: word[:4] != 'http', words)) / len(words)

df['avg_word_len'] = df['text'].apply(avg_word_len)

df['avg_word_len'].describe()

# Examine comments with shortest average word lengths.
for text in df.sort_values(by='avg_word_len')['text'].head(10):
    print(text, '\n')

Excluding links, these comments have no words at all.

# Examine comments with longest average word lengths.
for text in df.sort_values(by='avg_word_len', ascending=False)['text'].head(10):
    print(text, '\n')

df.head()

candidates = ['hour_posted', 'text_len', 'avg_word_len', 'pct_caps', 
              'tags_per_char', 
              'v_neg', 'v_neu', 'v_pos', 'v_compound', 
              'tb_polarity', 'tb_subjectivity', 'tb_nb_prob_neg', 
              'pc_prob_offensive']

X = df[candidates]
y = df['papi_toxicity']

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.8, random_state=0)

scaler = MinMaxScaler()

%%capture
X_train[candidates] = scaler.fit_transform(X_train)

X_train.describe()

%%capture
X_test[candidates] = scaler.transform(X_test)

## F-scores

best = SelectKBest(score_func=f_regression, k='all')

fit = best.fit(X_train, y_train)

dfscores = pd.DataFrame(fit.scores_)
dfcolumns = pd.DataFrame(X_train.columns)

#concat two dataframes for better visualization 
featureScores = pd.concat([dfcolumns,dfscores],axis=1)
featureScores.columns = ['Feature', 'F-Score']  #naming the dataframe columns
featureScores.sort_values('F-Score', ascending=False)

y_train.shape

## Recursive Feature Elimination

model = LinearRegression()

rfe = RFE(model, 5)
fit = rfe.fit(X_train, y_train)
print("Num Features: %s" % (fit.n_features_))
print("Selected Features: %s" % (fit.support_))
print("Feature Ranking: %s" % (fit.ranking_))
X_train.columns[fit.support_]

rfe = RFE(model, 4)
fit = rfe.fit(X_train, y_train)
print("Num Features: %s" % (fit.n_features_))
print("Selected Features: %s" % (fit.support_))
print("Feature Ranking: %s" % (fit.ranking_))
X_train.columns[fit.support_]

rfe = RFE(model, 3)
fit = rfe.fit(X_train, y_train)
print("Num Features: %s" % (fit.n_features_))
print("Selected Features: %s" % (fit.support_))
print("Feature Ranking: %s" % (fit.ranking_))
X_train.columns[fit.support_]

features = ['text_len', 'v_neg', 'pc_prob_offensive']

## Linear Regression

model.fit(X_train[features], y_train)

pred_train = model.predict(X_train[features])

pd.DataFrame(pred_train).describe()

pred_test = model.predict(X_test[features])

print('Train MAE: ', mean_absolute_error(pred_train, y_train))
print('Test MAE: ', mean_absolute_error(pred_test, y_test))

# Is that a lot of error or a little, given the scale and
# variance of our target variable?
y_train.describe()

# What about compared to a baseline baseline model, just predicting 
# the mean toxicity?
mean_pred_train = np.full_like(y_train, y_train.mean())
mean_pred_test = np.full_like(y_test, y_train.mean())

print('Train MAE (predicting mean): ', mean_absolute_error(mean_pred_train, y_train))
print('Test MAE (predicting mean): ', mean_absolute_error(mean_pred_test, y_test))

## Logistic Regression

# Convert PerspectiveAPI toxicity scores to binary class labels.
y_train_binary = (y_train > 0.7).astype(int)
y_test_binary = (y_test > 0.7).astype(int)

# Examine class balance in training dataset.
y_train_binary.value_counts(normalize=True)

model = LogisticRegression(class_weight='balanced')

model.fit(X_train[features], y_train_binary)

pred_train = model.predict(X_train[features])
pred_test = model.predict(X_test[features])

np.bincount(pred_train)

# Calculate metrics on training dataset.
train_accuracy = accuracy_score(y_train_binary, pred_train)
train_recall = recall_score(y_train_binary, pred_train)
train_precision = precision_score(y_train_binary, pred_train)

# Calculate metrics on test dataset.
test_accuracy = accuracy_score(y_test_binary, pred_test)
test_recall = recall_score(y_test_binary, pred_test)
test_precision = precision_score(y_test_binary, pred_test)

# Display metrics for training dataset.
print(f'Train accuracy: {train_accuracy:.4f}')
print(f'Train recall: {train_recall:.4f}')
print(f'Train precision: {train_precision:.4f}', '\n')

# Display metrics for test dataset.
print(f'Test accuracy: {test_accuracy:.4f}')
print(f'Test recall: {test_recall:.4f}')
print(f'Test precision: {test_precision:.4f}')
