==> Notebooks/09 - Neural Networks (manual wordlist).ipynb <==
# Import libraries.
import numpy as np
import pandas as pd
import spacy
import string

from joblib import dump, load

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics import accuracy_score, mean_absolute_error, precision_score, recall_score
from sklearn.model_selection import train_test_split

from tensorflow import keras
from tensorflow.keras.callbacks import EarlyStopping

# Load data.
df = pd.read_csv('../data/first_week_oct_2015_comments_by_top_400_with_scores_and_features_v2.csv')

# Preview data.
df.head()

# Review column list.
df.columns

# Define target column.
y = df['papi_toxicity']

## Text processing with spacy

nlp = spacy.load('en_core_web_sm')

# Split data into train and test subsets.
X_train, X_test, y_train, y_test = train_test_split(df['text'], 
                                                    y, 
                                                    train_size=0.8, 
                                                    random_state=0)

# Convert comment text to lowercase, remove punctuation, and trim 
# whitespace.
simple_text_train = X_train.str.lower()
simple_text_train = simple_text_train.str.replace('[{}]'.format(string.punctuation), '')
simple_text_train = simple_text_train.str.replace('\s+', ' ', regex=True)
simple_text_train = simple_text_train.str.strip()

simple_text_test = X_test.str.lower()
simple_text_test = simple_text_test.str.replace('[{}]'.format(string.punctuation), '')
simple_text_test = simple_text_test.str.replace('\s+', ' ', regex=True)
simple_text_test = simple_text_test.str.strip()

# Check results.
for text in simple_text_train.head():
    print(text, '\n')

# Tokenize text.
tokens_train = simple_text_train.apply(nlp.tokenizer)
tokens_test = simple_text_test.apply(nlp.tokenizer)

# # Lemmatize; remove stop words.
# lemmas_train = tokens_train.apply(lambda x: [token.lemma_ for token in x if not token.is_stop])
# lemmas_test = tokens_test.apply(lambda x: [token.lemma_ for token in x if not token.is_stop])

# Lemmatize; leave stop words. Will feature selection pick up "you"?
lemmas_train = tokens_train.apply(lambda x: [token.lemma_ for token in x])
lemmas_test = tokens_test.apply(lambda x: [token.lemma_ for token in x])

# Check results.
lemmas_train.head()

# Instantiate vectorizer object.
tfidf = TfidfVectorizer(min_df=0.0001, max_df=0.9999)

# Create a vocabulary and get word counts per document.
dtm_train = tfidf.fit_transform(lemmas_train.astype(str))
dtm_test = tfidf.transform(lemmas_test.astype(str))

# Get feature names to use as dataframe column headers.
dtm_train_orig = pd.DataFrame(dtm_train.todense(), columns=tfidf.get_feature_names())
dtm_test_orig = pd.DataFrame(dtm_test.todense(), columns=tfidf.get_feature_names())

# Copy document term matrices.
dtm_train = dtm_train_orig.copy()
dtm_test = dtm_test_orig.copy()

# Preview feature matrix.
dtm_train.head()

dump(tfidf, 'tfidf.joblib')

## Manual feature (word) selection

# Convert PerspectiveAPI toxicity scores to binary class labels.
y_train_binary = (y_train > 0.7).astype(int)
y_test_binary = (y_test > 0.7).astype(int)

# Check target distribution.
np.bincount(y_train_binary)

wordlist = ['addict', 
            'agenda', 
            'alllivesmatter',
            'anal', 
            'analretentive', 
            'anticorporate',
            'antihomosexuality', 
            'antivaxxer', 
            'antivaxxers', 
            'arse', 
            'arsehole',
            'asshole',
            'assholes',
            'awesome',
            'awesomeness',
            'banal',
            'banality',
            'basement',
            'batshit',
            'bitch',
            'blacklivesmatter',
            'bloody',
            'blow',
            'blowjobs',
            'blowup',
            'boob',
            'brain',
            'braindead',
            'brainfart',
            'brainfuck',
            'brilliant', 
            'brilliantly',
            'bro',
            'bros',
            'broscience',
            'broke',
            'brokenness',
            'brothel',
            'bullshit',
            'bullshitfree',
            'bum',
            'bummer',
            'cargocult',
            'celebrity',
            'cheerleader',
            'circlejerk',
            'clown',
            'clusterfuck',
            'clusterfucks',
            'cock',
            'cockamamie',
            'comment',
            'communist',
            'corporate',
            'cowardly',
            'cowboy',
            'crap',
            'crappy',
            'crapshoot',
            'craptastic',
            'crapware',
            'crazy',
            'cryptofascist',
            'cuckold',
            'cuckoo',
            'cult',
            'damn',
            'democrat',
            'deserve',
            'dick',
            'dicked',
            'dickhead',
            'dicking',
            'dicks',
            'dildo',
            'dumb',
            'dumbly',
            'elite',
            'excrement',
            'fag',
            'fail',
            'failing',
            'fake', 
            'fakery',
            'false', 
            'falsely',
            'fart',
            'fascist',
            'feel',
            'fool',
            'foolish',
            'foolproof',
            'forcedvaxxers',
            'fox',
            'fuck',
            'fuckedupness',
            'fullblown',
            'garbage',
            'genius',
            'goddamn',
            'goddamned',
            'greed',
            'greedy',
            'gross',
            'harebrained',
            'hell',
            'hellish',
            'helluva',
            'heterosexual',
            'hitler',
            'hitlers',
            'homo',
            'homophobia',
            'homophobic',
            'homosexual',
            'hooker',
            'idiot',
            'idiotic',
            'injustice',
            'irrational', 
            'irrationally',
            'jerk',
            'justice',
            'kneejerk',
            'knob',
            'leader',
            'lesbian',
            'liar',
            'lol',
            'loser',
            'maga',
            'master',
            'medium',
            'mindblowing',
            'mirror',
            'monocultural',
            'monoculture',
            'moron',
            'motherfucker',
            'motherfuckers',
            'motherfucking',
            'multicultural',
            'nazi',
            'nazis',
            'neonazi',
            'omg',
            'overblown',
            'overlord',
            'oxymoron',
            'people',
            'pervert',
            'perverter',
            'piss',
            'politician',
            'poop',
            'prick',
            'pyramid',
            'racist',
            'redpill',
            'republican',
            'retard',
            'rich',
            'right',
            'sapiosexual',
            'scum',
            'scumbag',
            'scummy',
            'seriously',
            'sexist',
            'shit',
            'shite',
            'shithole',
            'shitload',
            'shitshow',
            'shitstorm',
            'shitton',
            'shitty',
            'sjw', 
            'sleeze',
            'slimeballtype',
            'slimed',
            'slow',
            'skinnysexual',
            'smartarse',
            'socialist',
            'stupid',
            'stupidity',
            'stupidly',
            'suck',
            'sucker',
            'suckiness',
            'superrich',
            'terrorist',
            'trash',
            'troll',
            'trump',
            'tumblrsjw',
            'turd',
            'ugly',
            'undeservedly',
            'unreasonable',
            'vaxx',
            'vaxxed',
            'wank',
            'warrior',
            'waste', 
            'wasteful', 
            'wasteland',
            'weak',
            'weakly',
            'weakness',
            'web',
            'whistleblowing',
            'whore',
            'woman',
            'worthless',
            'wtf',
            'you',
            'your',
            'zomg']
len(wordlist)

dump(wordlist, 'wordlist.joblib')

dtm_train = dtm_train[wordlist]
dtm_test = dtm_test[wordlist]

y_train_binary.value_counts(normalize=True)

model = keras.Sequential()

callbacks = [EarlyStopping(monitor='val_loss', patience=2)]

model.add(keras.layers.Dense(1, input_dim=220, activation='relu'))
model.add(keras.layers.Dense(1, activation='sigmoid'))

model.compile(loss='binary_crossentropy', optimizer='adam')

model.fit(dtm_train, y_train_binary, epochs=25, 
          class_weight = {0: 0.010406, 1: 0.989594},
          callbacks=callbacks,
          validation_split=0.2, 
          verbose=1)

nn_train_pred = model.predict(dtm_train)
nn_test_pred = model.predict(dtm_test)
pd.DataFrame(nn_train_pred).describe()

# Calculate metrics on training dataset.
train_accuracy = accuracy_score(y_train_binary, np.rint(nn_train_pred))
train_recall = recall_score(y_train_binary, np.rint(nn_train_pred))
train_precision = precision_score(y_train_binary, np.rint(nn_train_pred))

# Calculate metrics on test dataset.
test_accuracy = accuracy_score(y_test_binary, np.rint(nn_test_pred))
test_recall = recall_score(y_test_binary, np.rint(nn_test_pred))
test_precision = precision_score(y_test_binary, np.rint(nn_test_pred))

# Display metrics for training dataset.
print(f'Train accuracy: {train_accuracy:.4f}')
print(f'Train recall: {train_recall:.4f}')
print(f'Train precision: {train_precision:.4f}', '\n')

# Display metrics for test dataset.
print(f'Test accuracy: {test_accuracy:.4f}')
print(f'Test recall: {test_recall:.4f}')
print(f'Test precision: {test_precision:.4f}')

# Calculate test metrics for majority classifier (perdicts no toxic comments).
mc_test_accuracy = accuracy_score(y_test_binary, np.full_like(y_test, 0))
mc_test_recall = recall_score(y_test_binary, np.full_like(y_test, 0))
mc_test_precision = precision_score(y_test_binary, np.full_like(y_test, 0))

# Display metrics for test dataset.
print(f'Majority classifier test accuracy: {mc_test_accuracy:.4f}')
print(f'Majority classifier test recall: {mc_test_recall:.4f}')
print(f'Majority classifier test precision: {mc_test_precision:.4f} (undefined)')

# Display comments identified as toxic by custom neural network.
for text in X_test[np.rint(nn_test_pred).squeeze().astype(bool)]:
    print(text, '\n')

model.save('simple_nn.h5')
